// src/data/characters-generated.ts
export const CHARACTERS: Record<string, CharacterData> = {
// ...
  HuTao: {
    key: 'HuTao',
    name: 'Hu Tao',
    element: 'pyro',
    weaponType: 'polearm',
    rarity: 5,
    base: { hp: 1210.7164, atk: 8.2859, def: 68.2062 },
    curve: { hp: 'GROW_CURVE_HP_S5', atk: 'GROW_CURVE_ATTACK_S5', def: 'GROW_CURVE_HP_S5' },
    bonusStat: 'cd',
    ascension: [
      { maxLevel: 20, hp: 0, atk: 0, def: 0, bonus: 0 },
      { maxLevel: 40, hp: 1038.0798, atk: 7.103891, def: 58.482, bonus: 0 },
      { maxLevel: 50, hp: 1775.6628, atk: 12.151393, def: 100.035, bonus: 0.096 },
      { maxLevel: 60, hp: 2759.107, atk: 18.881393, def: 155.439, bonus: 0.192 },
      { maxLevel: 70, hp: 3496.69, atk: 23.928896, def: 196.992, bonus: 0.192 },
      { maxLevel: 80, hp: 4234.273, atk: 28.976398, def: 238.545, bonus: 0.288 },
      { maxLevel: 90, hp: 4971.856, atk: 34.0239, def: 280.098, bonus: 0.384 },
    ],
  },
// ...
};