// src/data/weapons-generated.ts
export const WEAPONS: Record<string, WeaponData> = {
// ...
  StaffOfHoma: {
    key: 'StaffOfHoma',
    name: 'Staff of Homa',
    weaponType: 'polearm',
    rarity: 5,
    base: { atk: 45.9364, substat: 0.144 },
    curve: { atk: 'GROW_CURVE_ATTACK_301', substat: 'GROW_CURVE_CRITICAL_301' },
    substatKey: 'cd',
    ascension: [
      { maxLevel: 20, atk: 0 },
      { maxLevel: 40, atk: 31.1 },
      { maxLevel: 50, atk: 62.2 },
      { maxLevel: 60, atk: 93.4 },
      { maxLevel: 70, atk: 124.5 },
      { maxLevel: 80, atk: 155.6 },
      { maxLevel: 90, atk: 186.7 },
    ],
  },
// ...
};